pyproject.toml
[tool.poetry]
name = "spacypdfreader"
version = "0.2.0"
description = "A PDF to text extraction pipeline component for spaCy."
authors = ["SamEdwardes <edwardes.s@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/SamEdwardes/spaCyPDFreader"
keywords = ["python", "spacy", "nlp", "pdf", "pdfs"]

[tool.poetry.dependencies]
python = "^3.7"
spacy = "^3.0.0"
rich = "^10.15.2"
"pdfminer.six" = "^20211012"
pytesseract = {version = "^0.3.8", optional = true}
Pillow = {version = "^8.4.0", optional = true}
pdf2image = {version = "^1.16.0", optional = true}

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
en-core-web-sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl"}
black = "^21.12b0"

[tool.poetry.extras]
pytesseract = ["pytesseract", "Pillow", "pdf2image"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
